refactor(agent_harness): extract headless in-memory adapters
